+++ LottoDairy/Domain/UseCase/Chart/Protocol/ChartInformationUseCase.swift
@@ -1,0 +1,14 @@
+//
+//  ChartInformationUseCase.swift
+//  LottoDairy
+//
+//  Created by Sunny on 2023/09/03.
+//
+
+import Combine
+
+protocol ChartInformationUseCase {
+    func makeRangeOfYear() -> AnyPublisher<[Int], Never>
+    func makeYearAndMonthOfToday() -> [Int]
+    func makeChartInformationComponents(year: Int, month: Int) -> AnyPublisher<[ChartInformationComponents], Never>
+}
